requirements.txt
apache-beam[gcp]==2.9.0
google-cloud-spanner==1.7.1
google-cloud-core==0.29.1